requirements.txt
Django==4.1.2
django-appconf==1.0.5
django-channels==0.7.0
django-cors-headers==4.2.0
django-crispy-forms==1.14.0
django-image-cropping==1.7
Pillow==9.3.0
psycopg2-binary==2.9.9
psycopg2==2.9.9
gunicorn==21.2.0
whitenoise===6.6.0